// backend/src/import/oikotie/dtos/oikotie-property-data.dto.ts
export class OikotiePropertyDataDto {
  url: string;
  debtFreePrice: number;
  debtShare?: number;
  apartmentSize: number;
  maintenanceFee: number;
  waterFee?: number;
  financingFee?: number;
  address?: string;
  city?: string;
  postalCode?: string;
  district?: string;
  buildingYear?: number;
  propertyType?: string;
  roomDescription?: string;
  roomCount?: number;
  condition?: string;
  energyClass?: string;
  defaultImageUrl?: string;
}